refactor(portable-text-editor): move components and slightly improve them
Move components on root that are internal to own .components folder. Slightly TS-doc and improve them too
